Run XPath tests on all exported MEI files automatically
