--- test/mocha/test-lines.js
@@ -1,11 +1,0 @@
-"use strict";
-
-const utils = require('./utils');
-
-const mei = utils.getTestMeiDom('lines.mei');
-
-describe("Lines", () => {
-  it("matches XPath tests", function() {
-    utils.assertXpathAnnotations(mei);
-  });
-});
